requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
langchain-qdrant>=0.1.4
qdrant-client>=1.11.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
pypdf>=5.0.0
tiktoken>=0.7.0
pydantic>=2.9.0
reportlab>=4.0.0